pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["infra", "strategies"]

[project]
name = "aurum"
version = "0.1.0"
description = "Self-evolving quantitative investment system for individual investors"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.30",
    "pandas>=2.0",
    "numpy>=1.24",
    "openai>=1.0",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
    "socksio>=1.0",
    "supabase>=2.0",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = ["matplotlib>=3.7"]
